requirements.txt
agate==1.7.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alembic==1.18.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.13.0
apache-airflow==2.8.0
apache-airflow-providers-common-sql==1.19.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.10.0
argcomplete==3.6.3
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.6
click==8.3.1
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.0
dbt-core==1.7.0
dbt-duckdb==1.7.0
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
duckdb==0.10.0
email-validator==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
google-re2==1.1.20251105
googleapis-common-protos==1.73.1
graphviz==0.21
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==6.11.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
leather==0.4.1
limits==5.8.0
linkify-it-py==2.1.0
lockfile==0.12.2
Logbook==1.5.3
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
mashumaro==3.20
mdit-py-plugins==0.5.0
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.8.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
narwhals==2.18.1
networkx==3.6.1
numpy==1.26.4
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.0
parsedatetime==2.6
pathlib_abc==0.5.2
pathspec==0.11.2
pendulum==2.1.2
pillow==10.4.0
plotly==5.18.0
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
protobuf==4.25.9
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
PyJWT==2.12.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
pytzdata==2020.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.2
rpds-py==0.30.0
setproctitle==1.3.7
six==1.17.0
smmap==5.0.3
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.4.4
streamlit==1.31.0
tabulate==0.10.0
tenacity==8.5.0
termcolor==3.3.0
text-unidecode==1.3
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uc-micro-py==2.0.0
unicodecsv==0.14.1
universal_pathlib==0.3.10
urllib3==1.26.20
validators==0.35.0
Werkzeug==2.2.3
wrapt==2.1.2
WTForms==3.2.1
yarl==1.23.0
zipp==3.23.0
